code2
Exercise 1:
A. Write a function multiply that takes two integers as parameters and returns their
product.
B. Write an overloaded version of the function multiply that takes two doubles.
C. Write an overloaded version of the function multiply that take 3 inputs including one
integer, one double, and one flag value.
When the flag value is set false, the output result = integer * double.
When the flag value is set true, the output result = integer * double and the result is
ceiled or floored to generate the final integer value
The main will call the multiply functions with different inputs. Print out the final result values in
different cases